+++ langstream-api-gateway/src/main/java/ai/langstream/apigateway/websocket/handlers/AbstractHandler.java
@@ -25,9 +25,9 @@ import ai.langstream.api.runner.code.SimpleRecord;
 import ai.langstream.api.runner.topics.TopicConnectionsRuntime;
 import ai.langstream.api.runner.topics.TopicConnectionsRuntimeRegistry;
 import ai.langstream.api.runner.topics.TopicProducer;
-import ai.langstream.api.runner.topics.events.EventRecord;
-import ai.langstream.api.runner.topics.events.EventSources;
-import ai.langstream.api.runner.topics.events.GatewayEventData;
+import ai.langstream.api.events.EventRecord;
+import ai.langstream.api.events.EventSources;
+import ai.langstream.api.events.GatewayEventData;
 import ai.langstream.api.storage.ApplicationStore;
 import ai.langstream.apigateway.websocket.AuthenticatedGatewayRequestContext;
 import ai.langstream.impl.common.ApplicationPlaceholderResolver;
