+++ langstream-api-gateway/src/test/java/ai/langstream/apigateway/websocket/handlers/ProduceConsumeHandlerTest.java
@@ -30,9 +30,9 @@ import ai.langstream.api.model.Gateway;
 import ai.langstream.api.model.Gateways;
 import ai.langstream.api.model.StoredApplication;
 import ai.langstream.api.model.StreamingCluster;
-import ai.langstream.api.runner.topics.events.EventRecord;
-import ai.langstream.api.runner.topics.events.EventSources;
-import ai.langstream.api.runner.topics.events.GatewayEventData;
+import ai.langstream.api.events.EventRecord;
+import ai.langstream.api.events.EventSources;
+import ai.langstream.api.events.GatewayEventData;
 import ai.langstream.api.runtime.ClusterRuntimeRegistry;
 import ai.langstream.api.runtime.PluginsRegistry;
 import ai.langstream.api.storage.ApplicationStore;
