+++ langstream-api/src/main/java/ai/langstream/api/events/GatewayEventData.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ai.langstream.api.runner.topics.events;
+package ai.langstream.api.events;
 
 import java.util.Map;
 import lombok.AllArgsConstructor;
